add basic support for transation lifecycle hooks
Because you hand over control of the transaction creation to this
library, there is no way for you to know whether or not the current
transaction was sucessfully persisted to the database.

To circumvent that, we expose three helper methods that allow you to
hook into the transaction lifecycle and take appropriate action after a
commit/rollback.

  - `runOnTransactionCommit(cb)` takes a callback to be executed after
    the current transaction was sucessfully committed
  - `runOnTransactionRollback(cb)` takes a callback to be executed after
    the current transaction rolls back. The callback gets the error that
    initiated the roolback as a parameter.
  - `runOnTransactionEnd(cb)` takes a callback to be executed at the end
    of the current transactional context. If there was an error, it gets
    passed as an argument.